requirements.txt
agents==0.1.0
qdrant-client==1.9.1
cohere==4.3.1
python-dotenv==1.0.0
asyncio
logging
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pytest==7.4.3